requirements-prod.txt
# auditly Production Dependencies
# Core CLI framework
typer==0.12.3

# Configuration and validation
pydantic==2.6.4
PyYAML==6.0.2

# Cloud SDKs
boto3==1.34.34
google-cloud-iam==2.16.0
google-cloud-compute==1.14.1
google-cloud-storage==2.10.0
google-cloud-sql==0.4.0
google-cloud-logging==3.9.0

# Storage backends
minio==7.2.7
requests==2.32.3

# Security and cryptography
PyNaCl==1.5.0

# WebAssembly support for policy evaluation
wasmtime==25.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg>=3.1.18
aiosqlite==0.20.0

# Task scheduling
apscheduler==3.10.4

# Web framework and server
fastapi==0.109.2
uvicorn==0.27.1

# Rich text and formatting
rich==13.7.1

# Templating
jinja2==3.1.4
